refactor(tasks): consolidate one-off jobs into observing-task-runner
Replaces the two standalone job crates (backfill-occurrences,
replay-failed-records) with a single `observing-task-runner` binary that
dispatches them as subcommands. A new one-off task is now a module + one `Task`
arm — not a fresh crate, Dockerfile stage, CI matrix entry, and Cloud Run Job
each time.

"task-runner" (not "jobs"/"worker") to avoid the background-queue connotation:
these are operator-triggered, run-to-completion-and-exit tasks.

- crates/observing-task-runner: clap subcommand per task; each task's
  per-row logic lives in its module (backfill_occurrences, replay_failed_records)
  as `pub async fn run(Args)`. Shared tracing init in main. Behavior and SQL are
  unchanged from the merged binaries, so the .sqlx cache still matches.
- observing-bootstrap::job::connect_pool now resolves the connection via
  pg-url-env (DATABASE_URL, else DB_HOST/DB_NAME/DB_USER/DB_PASSWORD), so the
  runner can run under a least-privilege DB role in prod without a full URL.
- Dockerfile: one runtime-observing-task-runner stage (ENTRYPOINT so Cloud Run
  --args append the subcommand).
- CI: one build-matrix entry; a deploy-only Cloud Run Job (ingester role via
  DB_* + observing-db-ingester-password) — NOT auto-executed. Operators trigger
  runs with `gcloud run jobs execute observing-task-runner --args=<subcommand>,...`.

Heavyweight/odd-dep tasks should stay separate binaries rather than bloat this.

Verified: offline workspace build, fmt --check, clippy --workspace -D warnings,
sqlx prepare --check, and dry-run smoke tests of both subcommands.
