--- crates/backfill-occurrences/Cargo.toml
@@ -1,24 +1,0 @@
-[package]
-name = "backfill-occurrences"
-version = "0.1.0"
-edition = "2021"
-description = "One-shot CLI that re-fetches occurrence records from their authoring PDS and re-runs the upsert, to backfill newly-extracted columns (organismQuantity/organismQuantityType)"
-license = "MIT OR Apache-2.0"
-
-[dependencies]
-# `processing::occurrence_from_json` (feature-gated) + the occurrences upsert.
-observing-db = { path = "../observing-db", features = ["processing"] }
-# DID -> PDS resolution + `getRecord` (via fetch_record_by_aturi).
-atproto-blob-resolver = { path = "../atproto-blob-resolver" }
-# Shared batch-job scaffolding: CLI flags, pool setup, the drive loop.
-observing-bootstrap = { path = "../observing-bootstrap", default-features = false, features = [
-    "job",
-] }
-
-tokio = { workspace = true }
-sqlx = { workspace = true }
-chrono = { workspace = true }
-reqwest = { workspace = true }
-tracing = { workspace = true }
-tracing-subscriber = { workspace = true }
-clap = { version = "4", features = ["derive", "env"] }
